# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The api to Roamio's server.",
        "title": "Roamio",
        "contact": {},
        "version": "0.2"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/users/follow": {
            "post": {
                "description": "Add a follow relationship between a follower and target.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Follow a user or page",
                "parameters": [
                    {
                        "description": "Follower ID, Target ID, and type",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"Successfully followed\"}"
                    },
                    "400": {
                        "description": "{\"message\":\"Invalid request body or type\"}"
                    },
                    "404": {
                        "description": "{\"message\":\"User or target not found\"}"
                    },
                    "409": {
                        "description": "{\"message\":\"Cannot follow self or already following\"}"
                    },
                    "500": {
                        "description": "{\"message\":\"Internal server error\"}"
                    }
                }
            }
        },
        "/users/follow/check": {
            "post": {
                "description": "Check if a follower is following a target (user or page).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Check follow status",
                "parameters": [
                    {
                        "description": "Follower ID, Target ID, and type",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.IsFollowingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"true\"}"
                    },
                    "400": {
                        "description": "{\"message\":\"Invalid request body or type\"}"
                    },
                    "500": {
                        "description": "{\"message\":\"Database error\"}"
                    }
                }
            }
        },
        "/users/followers": {
            "post": {
                "description": "Get a list of followers for a user or page.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Retrieve followers",
                "parameters": [
                    {
                        "description": "Target ID and type",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of followers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Follower"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"message\":\"Invalid request body or type}"
                    },
                    "404": {
                        "description": "{\"message\":\"Target not found}"
                    },
                    "500": {
                        "description": "{\"message\":\"Failed to retrieve followers}"
                    }
                }
            }
        },
        "/users/followings": {
            "post": {
                "description": "Get a list of users or pages that a user is following.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Retrieve followings",
                "parameters": [
                    {
                        "description": "User ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of followings",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Following"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"message\":\"Invalid request body\"}"
                    },
                    "404": {
                        "description": "{\"message\":\"User not found\"}"
                    },
                    "500": {
                        "description": "{\"message\":\"Failed to retrieve followings\"}"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Logs in the user with a username or email and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.loginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/models.loginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid credentials"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Creates a new user with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.createUserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "409": {
                        "description": "Username or Email already exists"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/users/unfollow": {
            "post": {
                "description": "Remove a follow relationship between a follower and target.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Unfollow a user or page",
                "parameters": [
                    {
                        "description": "Follower ID, Target ID, and type",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\":\"Successfully unfollowed\"}"
                    },
                    "400": {
                        "description": "{\"message\":\"Invalid request body or type\"}"
                    },
                    "404": {
                        "description": "{\"message\":\"Follow relationship not found\"}"
                    },
                    "500": {
                        "description": "{\"message\":\"Failed to unfollow\"}"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Follower": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.Following": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Jane Smith"
                },
                "id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "models.GetFollowersRequest": {
            "type": "object",
            "required": [
                "target_id",
                "type"
            ],
            "properties": {
                "target_id": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "user",
                        "page"
                    ],
                    "example": "user"
                }
            }
        },
        "models.GetFollowingsRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.IsFollowingRequest": {
            "type": "object",
            "required": [
                "follower_id",
                "target_id",
                "type"
            ],
            "properties": {
                "follower_id": {
                    "type": "integer",
                    "example": 1
                },
                "target_id": {
                    "type": "integer",
                    "example": 2
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "user",
                        "page"
                    ],
                    "example": "user"
                }
            }
        },
...
        "models.UnfollowRequest": {
            "type": "object",
            "required": [
                "follower_id",
                "target_id",
                "type"
            ],
            "properties": {
                "follower_id": {
                    "type": "integer",
                    "example": 1
                },
                "target_id": {
                    "type": "integer",
                    "example": 2
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "user",
                        "page"
                    ],
                    "example": "user"
                }
            }
        },
        "models.createUserInput": {
            "type": "object",
            "properties": {
                "DOB": {
                    "description": "Date of birth of the user in YYYY-MM-DD format\nExample: 2000-04-08",
                    "type": "string"
                },
                "Email": {
                    "description": "Email address of the user\nExample: ppatel@ufl.edu",
                    "type": "string"
                },
                "Fullname": {
                    "description": "Full name of the user\nExample: Parth Patel",
                    "type": "string"
                },
                "Location": {
                    "description": "Location of the user\nExample: Kansas",
                    "type": "string"
                },
                "Password": {
                    "description": "Password for the user account\nExample: test123",
                    "type": "string"
                },
                "Username": {
                    "description": "Username for the user account\nExample: ParthPatel",
                    "type": "string"
                }
            }
        },
        "models.loginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "user": {
                    "type": "object",
                    "properties": {
                        "DOB": {
                            "type": "string",
                            "example": "2002-05-26"
                        },
                        "Email": {
                            "type": "string",
                            "example": "nsoni@ufl.edu"
                        },
                        "Fullname": {
                            "type": "string",
                            "example": "Nirvisha Soni"
                        },
                        "ID": {
                            "type": "integer",
                            "example": 1
                        },
                        "Location": {
                            "type": "string",
                            "example": "Gainesville"
                        },
                        "Password": {
                            "type": "string",
                            "example": ""
                        },
                        "Username": {
                            "type": "string",
                            "example": "NirvishaSoni"
                        }
                    }
                }
            }
        },
        "models.loginUser": {
            "type": "object",
            "required": [
                "password",
                "username_or_email"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username_or_email": {
                    "type": "string"
                }
            }
        },
...
    }
}